Move 'LocalMediaLoader' (and <MediaStore> collaborator) to it's own file in production
